Make "ResetPasswordURL" public without circ deps
