+++ cmd/frontend/auth/auth.go
@@ -4,11 +4,9 @@ package auth
 import (
 	"math/rand"
 	"net/http"
-	"strings"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/suspiciousnames"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/auth/userpasswd"
 	"github.com/sourcegraph/sourcegraph/internal/lazyregexp"
-	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
 // AuthURLPrefix is the URL path prefix under which to attach authentication handlers
@@ -59,52 +57,10 @@ func composeMiddleware(middlewares ...*Middleware) *Middleware {
 }
 
 // NormalizeUsername normalizes a proposed username into a format that meets Sourcegraph's
-// username formatting rules (based on, but not identical to
-// https://web.archive.org/web/20180215000330/https://help.github.com/enterprise/2.11/admin/guides/user-management/using-ldap):
-//
-// - Any characters not in `[a-zA-Z0-9-._]` are replaced with `-`
-// - Usernames with exactly one `@` character are interpreted as an email address, so the username will be extracted by truncating at the `@` character.
-// - Usernames with two or more `@` characters are not considered an email address, so the `@` will be treated as a non-standard character and be replaced with `-`
-// - Usernames with consecutive `-` or `.` characters are not allowed, so they are replaced with a single `-` or `.`
-// - Usernames that start with `.` or `-` are not allowed, starting periods and dashes are removed
-// - Usernames that end with `.` are not allowed, ending periods are removed
-//
-// Usernames that could not be converted return an error.
-//
-// Note: Do not forget to change database constraints on "users" and "orgs" tables.
+// username formatting rules.
 func NormalizeUsername(name string) (string, error) {
-	origName := name
-
-	// If the username is an email address, extract the username part.
-	if i := strings.Index(name, "@"); i != -1 && i == strings.LastIndex(name, "@") {
-		name = name[:i]
-	}
-
-	// Replace all non-alphanumeric characters with a dash.
-	name = disallowedCharacter.ReplaceAllString(name, "-")
-
-	// Replace all consecutive dashes and periods with a single dash.
-	name = consecutivePeriodsDashes.ReplaceAllString(name, "-")
-
-	// Trim leading and trailing dashes and periods.
-	name = sequencesToTrim.ReplaceAllString(name, "")
-
-	if name == "" {
-		return "", errors.Errorf("username %q could not be normalized to acceptable format", origName)
-	}
-
-	if err := suspiciousnames.CheckNameAllowedForUserOrOrganization(name); err != nil {
-		return "", err
-	}
-
-	return name, nil
+	return userpasswd.NormalizeUsername(name)
 }
-
-var (
-	disallowedCharacter      = lazyregexp.New(`[^\w\-\.]`)
-	consecutivePeriodsDashes = lazyregexp.New(`[\-\.]{2,}`)
-	sequencesToTrim          = lazyregexp.New(`(^[\-\.])|(\.$)|`)
-)
 
 // AddRandomSuffix appends a random 5-character lowercase alphabetical suffix (like "-lbwwt")
 // to the username to avoid collisions. If the username already ends with a dash, it is not
